skills/korean-multiturn-rag/benchmarks/behavioral/requirements.txt
anthropic>=0.40
pyyaml>=6.0
pytest>=8.0
